Uses salt lgpo module to write registry.pol
